src/test/java/com/leetcode/algorithm/easy/repeatedsubstringpattern/SolutionTest.java
package com.leetcode.algorithm.easy.repeatedsubstringpattern;

import org.junit.jupiter.api.Test;

import java.util.Arrays;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertEquals;

public class SolutionTest {
  @Test
  public void testCase1() {
    Solution solution = new Solution();
    assertEquals(true, solution.repeatedSubstringPattern("abab"));
  }

  @Test
  public void testCase2() {
    Solution solution = new Solution();
    assertEquals(false, solution.repeatedSubstringPattern("aba"));
  }

  @Test
  public void testCase3() {
    Solution solution = new Solution();
    assertEquals(true, solution.repeatedSubstringPattern("abcabcabcabc"));
  }
}
